+++ lib/src/widgets/switch.dart
@@ -44,11 +44,13 @@ class ShokoUISwitch extends StatefulWidget {
   final bool? enableOutline;
   final double? insidePadding;
 
+  final List<BoxShadow>? thumbShadow;
+
   const ShokoUISwitch({super.key,
     required this.value, this.onChange, this.width = 64,
     this.thumbColor, this.backgroundColor, this.outlineColor, this.inactiveThumbColor, this.inactiveBackgroundColor, this.inactiveOutlineColor,
     this.switchType,
-    this.enableOutline, this.insidePadding
+    this.enableOutline, this.insidePadding, this.thumbShadow
   });
 
   @override
@@ -123,10 +125,9 @@ class _ShokoUISwitchState extends State<ShokoUISwitch> {
           ),
           decoration: BoxDecoration(
             color: thumbColor,
+            boxShadow: widget.thumbShadow ?? context.shokoTheme?.switchTheme?.thumbShadow,
             borderRadius: BorderRadius.circular(100)
           ),
-          // height: 32,
-          // width: 32,
           duration: const Duration(milliseconds: 250),
           curve: Curves.easeInSine,
           )